+++ enviafib-model/src/main/java/es/caib/enviafib/model/entity/Peticio.java
@@ -23,8 +23,8 @@ public interface Peticio extends org.fundaciobit.genapp.common.IGenAppEntity {
 	public java.lang.String getDestinatarinif();
 	public void setDestinatarinif(java.lang.String _destinatarinif_);
 
-	public short getEstat();
-	public void setEstat(short _estat_);
+	public long getEstat();
+	public void setEstat(long _estat_);
 
 	public java.lang.Long getFitxerFirmatID();
 	public void setFitxerFirmatID(java.lang.Long _fitxerFirmatID_);
