+++ enviafib-persistence/src/main/java/es/caib/enviafib/persistence/PeticioJPAManager.java
@@ -58,7 +58,7 @@ public class PeticioJPAManager
         return list.toArray(new Peticio[list.size()]);
     };
 
-    public synchronized Peticio create( long _titolID_, java.sql.Timestamp _datacreacio_, long _fitxerID_, long _solicitantID_, java.lang.String _idiomaID_, java.lang.String _destinatarinif_, short _estat_, java.lang.Long _fitxerFirmatID_, java.lang.Long _peticioPortafib_) throws I18NException {
+    public synchronized Peticio create( long _titolID_, java.sql.Timestamp _datacreacio_, long _fitxerID_, long _solicitantID_, java.lang.String _idiomaID_, java.lang.String _destinatarinif_, long _estat_, java.lang.Long _fitxerFirmatID_, java.lang.Long _peticioPortafib_) throws I18NException {
         PeticioJPA __bean =  new PeticioJPA(_titolID_,_datacreacio_,_fitxerID_,_solicitantID_,_idiomaID_,_destinatarinif_,_estat_,_fitxerFirmatID_,_peticioPortafib_);
         return create(__bean);
     }
